+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * Stable ABI builds
 * Copy and `xor` optimized
 
+### Added
+* `toarrays` method
+
 ## [1.6.1](https://pypi.org/project/spector/1.6.1/) - 2025-10-27
 ### Added
 * Python 3.14 wheels